+ version bump + changes to "visibility and ignore" - now allows more per contact settings. + option restore last active view mode + removed old icon code. IcoLib API now required + removed old font configuration page - font service now required (as it is integrated in the core)
git-svn-id: http://miranda.googlecode.com/svn/trunk@4642 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/clist_nicer/clcmsgs.c
@@ -128,7 +128,7 @@ LRESULT ProcessExternalMessages(HWND hwnd, struct ClcData *dat, UINT msg, WPARAM
 			if(contact->type != CLCIT_CONTACT)
 				return 0;
 			contact->flags ^= CONTACTF_PRIORITY;
-			DBWriteContactSettingByte(contact->hContact, "CList", "Priority", contact->flags & CONTACTF_PRIORITY ? 1 : 0);
+			DBWriteContactSettingByte(contact->hContact, "CList", "Priority", (BYTE)(contact->flags & CONTACTF_PRIORITY ? 1 : 0));
 			pcli->pfnClcBroadcast(CLM_AUTOREBUILD, 0, 0);
 			return 0;
 		}
